refactor: merge activity viewmodel
  * merge ListActivityViewModel.kt into ListActivitiesViewModel.kt
