+++ app/src/main/java/com/android/sample/model/activity/ListActivitiesViewModel.kt
@@ -1,10 +1,13 @@
 package com.android.sample.model.activity
 
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewModelScope
 import com.google.firebase.Firebase
 import com.google.firebase.firestore.firestore
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 
 class ListActivitiesViewModel(private val repository: ActivitiesRepository) : ViewModel() {
@@ -28,16 +31,29 @@ class ListActivitiesViewModel(private val repository: ActivitiesRepository) : Vi
     }
   }
 
+  init {
+    repository.init { getActivities() }
+  }
+
   fun getNewUid(): String {
     return repository.getNewUid()
   }
 
-  // Initialize the repository
-  fun init(onSuccess: () -> Unit) {
-    repository.init(onSuccess)
+  fun addActivity(activity: Activity) {
+    repository.addActivity(activity, { getActivities() }, {})
   }
 
-  // Get all Activity items
+  fun updateActivity(activity: Activity) {
+    repository.updateActivity(activity, { getActivities() }, {})
+  }
+
+  fun deleteActivityById(id: String) {
+    repository.deleteActivityById(id, { getActivities() }, {})
+  }
+
+  fun selectActivity(activity: Activity) {
+    selectedActivity_.value = activity
+  }
 
   fun getActivities(onSuccess: () -> Unit = {}, onFailure: (Exception) -> Unit = {}) {
